[IMP] runbot: add authors and teams on bundles
With this commit, the authors involved in a bundle are computed.
Authors are found based on github logins first, then from the commit
authors and commiters and finally from the ngram extracted from the
bundle name.

The teams are infered from the authors found.

Finally an `Owning team` is automatically choosen (the first team of the teams)
or can be manually set.